Avoid unnecessary file-exists-p in maybe-mkdir.
mkdir doesn't signal errors if the directory exists and its optional second
arg is non-nil, so the check is unnecessary.

It also suffers from the fact that it's not atomic, so someone could create
the directory between the time we check and the time we try to create, however
highly improbable that may be.